# Conflicts:
#	src/main/scala/totoro/ocelot/brain/entity/InternetCard.scala
#	src/main/scala/totoro/ocelot/brain/entity/TextBuffer.scala
#	src/main/scala/totoro/ocelot/brain/entity/traits/GenericGPU.scala
#	src/main/scala/totoro/ocelot/brain/nbt/NBTTagCompound.java
#	src/main/scala/totoro/ocelot/brain/util/GenericTextBuffer.scala

+++ src/main/scala/totoro/ocelot/brain/nbt/NBTTagByte.java
@@ -50,7 +50,7 @@ public class NBTTagByte extends NBTNumber {
     }
 
     public long getLong() {
-        return (long) this.data;
+        return this.data;
     }
 
     public int getInt() {
@@ -58,7 +58,7 @@ public class NBTTagByte extends NBTNumber {
     }
 
     public short getShort() {
-        return (short) this.data;
+        return this.data;
     }
 
     public byte getByte() {
@@ -66,10 +66,10 @@ public class NBTTagByte extends NBTNumber {
     }
 
     public double getDouble() {
-        return (double) this.data;
+        return this.data;
     }
 
     public float getFloat() {
-        return (float) this.data;
+        return this.data;
     }
 }
